Scope secure namespace flags to separate object and add it to commands which need it
