+++ cli/pkg/kctrl/cmd/core/namespace_flags.go
@@ -13,14 +13,8 @@ import (
 	"github.com/spf13/pflag"
 )
 
-var sharedNamespaces = []string{
-	"default",
-	"kube-public",
-}
-
 type NamespaceFlags struct {
-	Name                    string
-	AllowedSharedNamespaces bool
+	Name string
 }
 
 func (s *NamespaceFlags) Set(cmd *cobra.Command, flagsFactory FlagsFactory) {
@@ -32,7 +26,6 @@ func (s *NamespaceFlags) SetWithPackageCommandTreeOpts(cmd *cobra.Command, flags
 	namespaceEnvVariableKey := fmt.Sprintf("%s_NAMESPACE", strings.ToUpper(opts.BinaryName))
 	name := flagsFactory.NewNamespaceNameFlag(&s.Name, namespaceEnvVariableKey)
 	cmd.Flags().VarP(name, "namespace", "n", fmt.Sprintf("Specified namespace ($%s or default from kubeconfig)", namespaceEnvVariableKey))
-	cmd.Flags().BoolVar(&s.AllowedSharedNamespaces, "dangerous-allow-use-of-shared-namespace", false, "Allow use of shared namespaces (default: false)")
 }
 
 type NamespaceNameFlag struct {
@@ -64,20 +57,6 @@ func (s *NamespaceNameFlag) Resolve() error {
 
 	*s.value = value
 
-	return nil
-}
-
-func (s *NamespaceFlags) CheckForDisallowedSharedNamespaces() error {
-	if s.AllowedSharedNamespaces {
-		return nil
-	}
-	for _, ns := range sharedNamespaces {
-		if s.Name == ns {
-			return fmt.Errorf(`Creating sensitive resources in a shared namespace (%s)
-			(hint: Specify a namespace using the '-n' flag or use kubeconfig to change default namespace 'kubectl config set-context --current --namespace=private-namespace'.
-			Or use '--dangerous-allow-allow-use-of-shared-namespace=%s' to allow use of shared namespace)`, s.Name, s.Name)
-		}
-	}
 	return nil
 }
